C - Hello, World

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@
  */
 int main(void)
 {
-	printf("The size of char is %lu byte(s).\n", (unsigned long)sizeof(char));
-	printf("The size of int is %lu byte(s).\n", (unsigned long)sizeof(int));
-	printf("The size of float is %lu byte(s).\n", (unsigned long)sizeof(float));
-	printf("The size of long int is %lu byte(s).\n", (unsigned long)sizeof(long int));
-	printf("The size of long long int is %lu byte(s).\n", (unsigned long)sizeof(long long int));
+	printf("The size of char is %lu byte(s).\n", sizeof(char));
+	printf("The size of int is %lu byte(s).\n", sizeof(int));
+	printf("The size of float is %lu byte(s).\n", sizeof(float));
+	printf("The size of long int is %lu byte(s).\n", sizeof(long int));
+	printf("The size of long long int is %lu byte(s).\n", sizeof(long long int));
 	return (0);
 }
